Refactor: rename dummy subclases
